+++ netuitive/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 __author__ = 'Netuitive, Inc'
-__version__ = '0.0.3'
+__version__ = '0.0.6'
 
 from .client import Client
 from .element import Element